# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YoYo Map API — global yo-yo player community map
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/update:
    post:
      operationId: updateProfile
      tags: [profile]
      summary: Update a profile entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
        "400":
          description: Bad request or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /profile/delete:
    post:
      operationId: deleteProfile
      tags: [profile]
      summary: Delete a profile entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required: [token]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
          required: [code, message]
      required: [error]
...
    UpdateProfileBody:
      type: object
      properties:
        token:
          type: string
        display_name:
          type: string
        bio:
          type: string
          nullable: true
        socials:
          type: object
          additionalProperties:
            type: string
        club_meeting_info:
          type: string
          nullable: true
        club_venue_public:
          type: boolean
        contact_name:
          type: string
          nullable: true
        address_line:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        hours:
          type: string
          nullable: true
      required: [token]
...